requirements.txt
streamlit
pypdf2
google-generativeai==0.3.2
python-dotenv==1.0.1
youtube_transcript_api